lib/idxd: match batch refcnt and index to match DSA spec
DSA 2.0 spec defines the size of Maximum Supported Batch Size
as 4 bits (see 9.2.2 General Capabilities Register).
Batch index and refcnt was just an uint8_t that couldn't fit
all possible values.

Right now the batch size is set to DESC_PER_BATCH,
but it could be configurable or taken from the aforementioned
General Cap register.

Reviewed-on: https://review.spdk.io/gerrit/c/spdk/spdk/+/21242
Community-CI: Mellanox Build Bot
Tested-by: SPDK CI Jenkins <sys_sgci@intel.com>

+++ lib/idxd/idxd.c
@@ -514,7 +514,7 @@ idxd_batch_cancel(struct spdk_idxd_io_channel *chan, int status)
 	batch = chan->batch;
 	assert(batch != NULL);
 
-	if (batch->index == UINT8_MAX) {
+	if (batch->index == UINT16_MAX) {
 		SPDK_ERRLOG("Cannot cancel batch, already submitted to HW.\n");
 		return -EINVAL;
 	}
@@ -581,7 +581,7 @@ idxd_batch_submit(struct spdk_idxd_io_channel *chan,
 			STAILQ_INSERT_TAIL(&chan->ops_outstanding, (struct idxd_ops *)&batch->user_ops[i],
 					   link);
 		}
-		batch->index = UINT8_MAX;
+		batch->index = UINT16_MAX;
 	}
 
 	chan->batch = NULL;
