+++ lib/idxd/idxd_internal.h
@@ -53,8 +53,8 @@ struct idxd_batch {
 	struct idxd_hw_desc		*user_desc;
 	struct idxd_ops			*user_ops;
 	uint64_t			user_desc_addr;
-	uint8_t				index;
-	uint8_t				refcnt;
+	uint16_t			index;
+	uint16_t			refcnt;
 	struct spdk_idxd_io_channel	*chan;
 	TAILQ_ENTRY(idxd_batch)		link;
 };
